+++ src/ext/BalExtension/wixstdba/precomp.h
@@ -48,6 +48,7 @@
 HRESULT CreateBootstrapperApplication(
     __in HMODULE hModule,
     __in BOOL fPrereq,
+    __in LPCWSTR wzFailedMessage,
     __in IBootstrapperEngine* pEngine,
     __in const BOOTSTRAPPER_COMMAND* pCommand,
     __out IBootstrapperApplication** ppApplication